% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plans.R
\name{get_project_plan}
\alias{get_project_plan}
\title{Get the Project Plan}
\usage{
get_project_plan()
}
\value{
a \code{drake} plan
}
\description{
Use \code{\link[drake]{drake_plan}} to create the project
plan.
}
